requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
chumpy==0.70
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.3
PyYAML==6.0.2
regex==2025.7.34
scipy==1.15.3
six==1.17.0
smplx==0.1.28
sympy==1.14.0
tensorboardX==2.6.4
tomli==2.2.1
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
wcwidth==0.2.13
yarl==1.20.1
